+++ src/main/java/com/tbsc/consultant/ConsultantRepository.java
@@ -7,10 +7,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ConsultantRepository extends JpaRepository<Consultant, Integer> {
+public interface ConsultantRepository extends JpaRepository<Consultant, Long> {
     Optional<Consultant> findByNum(Long num);
 
     Page<Consultant> findAll(Pageable pageable);
